More restructuring of ast.go
